StudyGroup Insert(regist) 코드 수정 완료
studygroupCreateDate 같은 회원이 직접 설정할 수 없는 필드의 값들을 변수로 미리 생성하고 setter와 getter를 사용해 insert 부분을 수정했습니다.
