+++ innerjoinus-studygroup-source/src/main/java/org/thisdote/innerjoinus/InnerjoinusStudygroupSourceApplication.java
@@ -4,9 +4,7 @@ import org.modelmapper.ModelMapper;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.context.annotation.Bean;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 
 @SpringBootApplication
 @EnableDiscoveryClient
@@ -19,11 +17,5 @@ public class InnerjoinusStudygroupSourceApplication {
     @Bean
     public ModelMapper getModelMapper() {
         return new ModelMapper();
-    }
-
-    /* 설명. 비밀번호 암호화를 위한 BCrypt bean 추가 */
-    @Bean
-    public BCryptPasswordEncoder bCryptPasswordEncoder() {
-        return new BCryptPasswordEncoder();
     }
 }
