+++ innerjoinus-studygroup-source/src/main/java/org/thisdote/innerjoinus/studygroup/command/repository/StudyGroupRepository.java
@@ -1,0 +1,8 @@
+package org.thisdote.innerjoinus.studygroup.command.repository;
+
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.thisdote.innerjoinus.studygroup.command.entity.StudyGroupEntity;
+
+public interface StudyGroupRepository extends JpaRepository<StudyGroupEntity, Integer> {
+}
