+++ innerjoinus-studygroup-source/src/main/java/org/thisdote/innerjoinus/studygroup/command/vo/RequestStudyGroup.java
@@ -1,0 +1,20 @@
+package org.thisdote.innerjoinus.studygroup.command.vo;
+
+import lombok.*;
+import org.springframework.stereotype.Component;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+@ToString
+@Component
+public class RequestStudyGroup {
+    private int studygroupType;
+    private java.util.Date studygroupCreateDate;
+    private int studygroupMemberCount;
+    private int studygroupActivationStatus;
+    private java.util.Date studygroupStudyTime;
+    private String studygroupContent;
+    private int studygroupDeleteStatus;
+}
