+++ innerjoinus-studygroup-source/src/main/java/org/thisdote/innerjoinus/studygroup/dto/StudyGroupDTO.java
@@ -1,0 +1,22 @@
+package org.thisdote.innerjoinus.studygroup.dto;
+
+import lombok.*;
+
+import java.util.Date;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Setter
+@Getter
+@ToString
+public class StudyGroupDTO {
+
+    private int studyGroupId;
+    private int studyGroupType;
+    private Date studyGroupCreatedDate;
+    private int studyGroupMemberCount;
+    private int studyGroupActivationStatus;
+    private Date studyGroupStudyTime;
+    private String studyGroupContent;
+    private int studyGroupDeleteStatus;
+}
