+++ innerjoinus-studygroup-source/src/main/java/org/thisdote/innerjoinus/studygroup/query/SearchStudyGroupTypeEnum.java
@@ -1,0 +1,5 @@
+package org.thisdote.innerjoinus.studygroup.query;
+
+public enum SearchStudyGroupTypeEnum {
+    CERTIFICATION, PROGRAMMING, JOB_SEARCH, ENGLISH, ETC
+}
