+++ innerjoinus-studygroup-source/src/main/java/org/thisdote/innerjoinus/studygroup/query/repository/StudyGroupMapper.java
@@ -1,0 +1,20 @@
+package org.thisdote.innerjoinus.studygroup.query.repository;
+
+import org.apache.ibatis.annotations.Mapper;
+import org.thisdote.innerjoinus.studygroup.dto.StudyGroupDTO;
+
+import java.util.List;
+
+@Mapper
+public interface StudyGroupMapper {
+
+    List<StudyGroupDTO> selectAllStudyGroup();
+
+    List<StudyGroupDTO> selectAllEnglishStudyGroup();
+
+    List<StudyGroupDTO> selectAllLicenseStudyGroup();
+
+    List<StudyGroupDTO> selectStudyGroupByStatus(int activationStatus);
+
+    List<StudyGroupDTO> selectStudyGroupsByType(int studyGroupType);
+}
