+++ innerjoinus-studygroup-source/src/main/java/org/thisdote/innerjoinus/studygroup/query/vo/ResponseStudyGroup.java
@@ -1,0 +1,18 @@
+package org.thisdote.innerjoinus.studygroup.query.vo;
+
+import lombok.*;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Setter
+@Getter
+@ToString
+public class ResponseStudyGroup {
+    private int studygroupType;
+    private String studygroupCreateDate;
+    private int studygroupMemberCount;
+    private int studygroupActivationStatus;
+    private String studygroupStudyTime;
+    private String studygroupContent;
+    private int studygroupDeleteStatus;
+}
