+++ innerjoinus-studygroup-source/src/main/java/org/thisdote/innerjoinus/studygroup/query/vo/StudyGroupVo.java
@@ -1,0 +1,4 @@
+package org.thisdote.innerjoinus.studygroup.query.vo;
+
+public class StudyGroupVo {
+}
