--- innerjoinus-studygroup-source/src/main/java/org/thisdote/innerjoinus/studygroupservice/client/StudyGroupClient.java
@@ -1,8 +1,0 @@
-package org.thisdote.innerjoinus.studygroupservice.client;
-
-import org.springframework.cloud.openfeign.FeignClient;
-
-@FeignClient(name="swcamp-innerjoinus-studygroup", url="localhost:8888")
-public interface StudyGroupClient {
-
-}
