--- innerjoinus-studygroup-source/src/main/java/org/thisdote/innerjoinus/studygroupservice/command/service/StudyGroupCommandService.java
@@ -1,16 +1,0 @@
-package org.thisdote.innerjoinus.studygroupservice.command.service;
-
-import org.springframework.stereotype.Service;
-import org.thisdote.innerjoinus.studygroupservice.dto.StudyGroupCommandDTO;
-
-
-@Service
-public interface StudyGroupCommandService {
-
-
-    void updateStudyGroup(StudyGroupCommandDTO studyGroupCommandDTO);
-
-    void removeStudyGroup(StudyGroupCommandDTO studyGroupCommandDTO);
-
-    void insertStudygroup(StudyGroupCommandDTO studyGroupCommandDTO);
-}
