--- innerjoinus-studygroup-source/src/main/java/org/thisdote/innerjoinus/studygroupservice/command/vo/RequestStudyGroup.java
@@ -1,20 +1,0 @@
-package org.thisdote.innerjoinus.studygroupservice.command.vo;
-
-import lombok.*;
-import org.springframework.stereotype.Component;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-@ToString
-@Component
-public class RequestStudyGroup {
-    private int studygroupType;
-    private java.util.Date studygroupCreateDate;
-    private int studygroupMemberCount;
-    private int studygroupActivationStatus;
-    private java.util.Date studygroupStudyTime;
-    private String studygroupContent;
-    private int studygroupDeleteStatus;
-}
