--- innerjoinus-studygroup-source/src/main/java/org/thisdote/innerjoinus/studygroupservice/dto/StudyGroupDTO.java
@@ -1,22 +1,0 @@
-package org.thisdote.innerjoinus.studygroupservice.dto;
-
-import lombok.*;
-
-import java.util.Date;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Setter
-@Getter
-@ToString
-public class StudyGroupDTO {
-
-    private int studyGroupId;
-    private int studyGroupType;
-    private Date studyGroupCreatedDate;
-    private int studyGroupMemberCount;
-    private int studyGroupActivationStatus;
-    private Date studyGroupStudyTime;
-    private String studyGroupContent;
-    private int studyGroupDeleteStatus;
-}
