--- innerjoinus-studygroup-source/src/main/java/org/thisdote/innerjoinus/studygroupservice/query/SearchStudyGroupTypeEnum.java
@@ -1,5 +1,0 @@
-package org.thisdote.innerjoinus.studygroupservice.query;
-
-public enum SearchStudyGroupTypeEnum {
-    CERTIFICATION, PROGRAMMING, JOB_SEARCH, ENGLISH, ETC
-}
