--- innerjoinus-studygroup-source/src/main/java/org/thisdote/innerjoinus/studygroupservice/query/vo/StudyGroupVo.java
@@ -1,4 +1,0 @@
-package org.thisdote.innerjoinus.studygroupservice.query.vo;
-
-public class StudyGroupVo {
-}
