+++ innerjoinus-studygroup-source/src/main/resources/application.yml
@@ -12,8 +12,6 @@ spring:
     password: swcamp
 
   jpa:
-    hibernate:
-      ddl-auto: create
     show-sql: true
     database: mysql
     properties:
